minio/requirements.txt
fastapi
uvicorn[standard]
minio
python-dotenv
